LayoutTests/http/tests/security/xss-DENIED-synchronous-form-expected.txt

CONSOLE MESSAGE: Unsafe JavaScript attempt to access frame with URL http://127.0.0.1:8000/security/xss-DENIED-synchronous-form.html from frame with URL http://localhost:8000/security/resources/iframe-for-synchronous-form.html. Domains, protocols and ports must match.

Waiting for iframe...
Done.
